PP9-6335: Adjust to new endpoints, move tests to appropriate test classes
