+++ src/Picturepark.SDK.V1.Contract/IPictureparkService.cs
@@ -28,8 +28,6 @@ namespace Picturepark.SDK.V1.Contract
 
         ISchemaTransferClient SchemaTransfer { get; }
 
-        IPublicAccessClient PublicAccess { get; }
-
         IShareClient Share { get; }
 
         ITransferClient Transfer { get; }
@@ -43,7 +41,5 @@ namespace Picturepark.SDK.V1.Contract
         IChannelClient Channel { get; }
 
         IServiceProviderClient ServiceProvider { get; }
-
-        IShareAccessClient ShareAccess { get; }
     }
 }
